requirements.txt
keyring
pytz
ruamel.yaml
click
pyrfc3339
pyotp
awscli
structlog
retrying
tzlocal
natural
requests_toolbelt
selenium
furl
tabulate
ruamel.yaml
colorama
marshmallow
marshmallow-dataclass
